src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
dj-database-url==2.3.0
    # via -r src/requirements/requirements.in
django==5.1.2
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-crispy-forms
    #   django-filter
    #   django-pwa
    #   django-redis
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   slippers
django-admin-interface==0.29.1
    # via -r src/requirements/requirements.in
django-allauth==65.1.0
    # via -r src/requirements/requirements.in
django-allauth-ui==1.5.0
    # via -r src/requirements/requirements.in
django-colorfield==0.11.0
    # via django-admin-interface
django-crispy-forms==2.3
    # via -r src/requirements/requirements.in
django-filter==24.3
    # via -r src/requirements/requirements.in
django-pwa==2.0.1
    # via -r src/requirements/requirements.in
django-redis==5.4.0
    # via -r src/requirements/requirements.in
django-widget-tweaks==1.5.0
    # via
    #   -r src/requirements/requirements.in
    #   django-allauth-ui
djangorestframework==3.15.2
    # via
    #   -r src/requirements/requirements.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r src/requirements/requirements.in
drf-spectacular-sidecar==2024.7.1
    # via -r src/requirements/requirements.in
gunicorn==23.0.0
    # via -r src/requirements/requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
packaging==24.1
    # via gunicorn
pillow==11.0.0
    # via
    #   -r src/requirements/requirements.in
    #   django-colorfield
psycopg2-binary==2.9.10
    # via -r src/requirements/requirements.in
python-decouple==3.8
    # via -r src/requirements/requirements.in
python-slugify==8.0.4
    # via django-admin-interface
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   slippers
redis==5.2.0
    # via django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r src/requirements/requirements.in
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
slippers==0.6.2
    # via
    #   -r src/requirements/requirements.in
    #   django-allauth-ui
sqlparse==0.5.1
    # via django
text-unidecode==1.3
    # via python-slugify
typeguard==2.13.3
    # via slippers
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   slippers
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via requests
whitenoise==6.7.0
    # via -r src/requirements/requirements.in
